requirements/production.txt
-r common.txt
django-auth-ldap==1.7.0
